+++ farben/src/macros/mod.rs
@@ -7,3 +7,4 @@ pub mod color;
 pub mod eprint;
 pub mod format;
 pub mod markdown;
+pub mod strip;